feat(tool): properly decode JWTs
The decoding of the json web tokens has always been a bit unstable, as
there are many corner cases to consider. Instead of fixing this again
and again, we just use a small jwt library. The library is only added as
a tool dependency, so libsso-mib users do not need to install it.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>