feat(credential-helper): allow to pass options via CLI
We currently require the user to specify the options via environment
variables. While this makes it easy to pass the data, it does not work
if different client ids are used, depending on the auth target. It also
is more complicated when globally enrolled. We now also support to pass
data via the CLI.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>