chore: make libjwt optional
We always inteded to have this a an optional tool-only dependency.
However, we missed to actually implement it like that.

We now make the dependency optional and print an error in case token
decoding is requested but the tool was not built with the library. We
further add a meson option "libjwt" to configure if the dependency is
optional.

Closes: #15
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>