chore: make code compile warning free
We globally disable the unused-argument warning, as this is hard to
avoid with conditionally compiled code and in general does not provide
much value. For generated code, we disable the pedantic warning level,
as this is out of our control anyways. To technically achieve this, we
move the generated code to a static library that is only consumed
internally.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>